refactor to shared ChatType
